refactor(PartialDateUi): begin splitting the file into smaller
