+++ specifyweb/frontend/js_src/lib/utils/parser/dateConfig.ts
@@ -1,1 +1,5 @@
+/**
+ * See also FormPlugins/useDatePreferences.tsx
+ */
+
 export const databaseDateFormat = 'YYYY-MM-DD';
